+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+## In development
+
+* Add method to reorder lists. Thanks @ZelKami! #131
+
 ## 6.11.2
 _2022-09-01_